+++ js/content.js
@@ -1,0 +1,1 @@
+/*Функционал по просмотру загруженных фотографий */
