+++ js/editor.js
@@ -1,0 +1,1 @@
+/*Функционал по редактированию фотографий*/
