+++ js/form.js
@@ -1,0 +1,1 @@
+/* Отправка формы на сервер */
